desktop/requirements.txt
pywinauto>=0.6.8
psutil>=5.9.0
pystray>=0.19.5
Pillow>=10.0.0
pynput>=1.7.7
pyautogui>=0.9.54
pytest>=8.0.0
tomli>=2.0.1; python_version < "3.11"
